+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## #.#.#
 
+## Changed
+
+- Separate security for file in another enum FilesSecurity
+
 ## Added
 
+- Add fileSecurity, security props to IFileItem
+- Add itemSecurity prop to IContextMenuItem
 - Add withoutBodyPadding and withoutHeaderMargin properties to IModalDialog
 - Add updateCreateDialogModal action
 - Add isAutoFocusOnError, errorText, onError, onChange properties to ICreateDialog